particle to cell mapping code restructuring
